+++ config/__init__.py
@@ -1,2 +1,5 @@
-﻿"""Django project package."""
+"""Django project package."""
 
+from .celery import app as celery_app
+
+__all__ = ("celery_app",)